requirements.txt
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
python-dateutil==2.7.2
python-twitter==3.4.1
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
